gestion erreur like / retrait logger http / Serializer dblogger / move and rename logs
